move exp and epx_grad kernel to activation_op_npu.cc, delete attrs
