--- paddle/fluid/operators/exp_op.cc
@@ -1,19 +1,0 @@
-/* Copyright (c) 2018 PaddlePaddle Authors. All Rights Reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License. */
-
-/*
-        This File can make gcc to compile './exp_op_npu.cc' .
-        If this file is deleted, exp_op_npu will not be compiled!
-        Refer to ./activation_op.cc to get ExpOP definition
-*/
